Documented %x adding trailing space to returns
While that might seem like a "bug-feature", I actually intended it
this way, I just didn't write that in the docs to begin with. I
wanted a way to keep Pyhoc from jamming everything together, so
I had it add a space to whatever it returned.

I intend to make three other versions for each substitution, so
I have one that returns `output `, `output`, `[output] `, and
`[output]`, so users can choose their own style of prompt